Update command list to apply NODE_NAME env var correctly
We had an issue when deploying the daemon set and found the `NODE_NAME` was not translated correctly through args. Consequently the network-metrics-daemon pods was unable to properly execute as the environment variable was nullified.

By utilizing a different command method, the pod is now able to retrieve the `NODE_NAME` env variable.

Running on 8 nodes, version `v1.29.4+rke2r1`.